### 2023-08-29 #5787 @editorialbot
[PRE REVIEW]: Umami: A Python toolkit for jet flavour tagging
**Submitting author:** @philippgadow (<a href="http://orcid.org/0000-0003-4475-6734">Philipp Gadow</a>)
**Repository:** https://github.com/umami-hep/umami
**Branch with paper.md** (empty if default branch): paper
**Version:** 0.20
