### 2023-09-11 #5833 @editorialbot
[REVIEW]: Umami: A Python toolkit for jet flavour tagging
**Submitting author:** @philippgadow (<a href="http://orcid.org/0000-0003-4475-6734">Philipp Gadow</a>)
**Repository:** https://github.com/umami-hep/umami
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.22
