### 2023-10-03 #5833 @matthewfeickert
Thanks @jpata. As those seem like actionable items can you maybe split those out into individual `[JOSS Review]: thing` Issues?
