### 2023-10-04 umami-hep/umami#2 @jpata
JOSS review: add clear example on how to use the JetClass dataset
Add some example configs and documentation on how to use the umami library with the open JetClass dataset (or some alternative open dataset).
