### 2023-10-09 #5833 @matthewfeickert
:wave: @philippgadow, as there hasn't been any response or action towards @jpata's review Issues can you confirm that you and the `umami` team have seen them?
