### 2023-10-09 #5833 @philippgadow
Dear @matthewfeickert , thank you for the ping!

And thank you very much @jpata for the kind words and your work in reviewing the software!
We'll discuss the comments in the flavour tagging group and will address them in the [issue](https://github.com/umami-hep/umami/issues/1) you opened.

Cheers,
Philipp for the umami developers team
