### 2023-10-12 #5833 @jpata
No news from me, I've left the review at https://github.com/umami-hep/umami/issues/1 (also made separate issues) and it's being worked on as I understand from https://github.com/openjournals/joss-reviews/issues/5833#issuecomment-1752803138.
