+++ umami/train_tools/nn_tools.py
@@ -344,7 +344,6 @@ def create_metadata_folder(
         # Change the paths for the preprocess config and var dict in the
         # train_config
         if file_path == train_config_path:
-
             replace_line_in_file(
                 new_file_path,
                 "preprocess_config:",
@@ -1005,7 +1004,10 @@ def load_validation_data(
 
         # Check which tagger is used and load the neede data
         if tagger.casefold() in ("dips", "dips_attention"):
-            (x_valid, y_valid,) = get_test_sample_trks(
+            (
+                x_valid,
+                y_valid,
+            ) = get_test_sample_trks(
                 input_file=val_file_config["path"],
                 var_dict=train_config.general.var_dict,
                 scale_dict=train_config.general.preprocess_config.general.dict_file,
@@ -1016,7 +1018,10 @@ def load_validation_data(
             )
 
         elif tagger.casefold() in ("dl1"):
-            (x_valid, y_valid,) = get_test_sample(
+            (
+                x_valid,
+                y_valid,
+            ) = get_test_sample(
                 input_file=val_file_config["path"],
                 var_dict=train_config.general.var_dict,
                 scale_dict=train_config.general.preprocess_config.general.dict_file,
@@ -1027,7 +1032,11 @@ def load_validation_data(
             )
 
         elif tagger.casefold() in ("umami", "umami_cond_att", "cads"):
-            (x_valid, x_valid_trk, y_valid,) = get_test_file(
+            (
+                x_valid,
+                x_valid_trk,
+                y_valid,
+            ) = get_test_file(
                 input_file=val_file_config["path"],
                 var_dict=train_config.general.var_dict,
                 scale_dict=train_config.general.preprocess_config.general.dict_file,
@@ -1220,7 +1229,13 @@ def evaluate_model_umami(
                 data_dict[f"X_valid_trk_{val_file_identifier}"],
                 data_dict[f"X_valid_{val_file_identifier}"],
             ]
-        (loss, dips_loss, umami_loss, dips_accuracy, umami_accuracy,) = model.evaluate(
+        (
+            loss,
+            dips_loss,
+            umami_loss,
+            dips_accuracy,
+            umami_accuracy,
+        ) = model.evaluate(
             x_valid,
             data_dict[f"Y_valid_{val_file_identifier}"],
             batch_size=batch_size,
@@ -1461,7 +1476,6 @@ def calc_validation_metrics(
 
     # Open the json file and load the training out
     try:
-
         # Get val dict file name
         _, val_output_file_path = get_metrics_file_name(
             working_point=eval_parameters.working_point,
