### 2023-10-23 #5833 @hqucms
@matthewfeickert @philippgadow I am very sorry for the delay -- this somehow slipped through the cracks...

I have reviewed the paper and the software documentation and updated my review checklist. It is clearly an impressive amount of work -- the software is well structured and easy-to-use, and the documentation is detailed and clear. Congratulations to all the developers!

My only comment actually concurs with @jpata's https://github.com/umami-hep/umami/issues/2, i.e., the lack of a self-contained example demonstrating the full workflow from preprocessing to training and then plotting, using a public dataset. Without that it is difficult to test the functionality of the software.
