### 2023-10-30 16cbb2e
Change truth label names
