+++ umami/configs/Umami_Variables.yaml
@@ -78,16 +78,16 @@ track_train_variables:
 
 track_truth_variables:
   tracks:
-    - truthOriginLabel
-    - truthVertexIndex
+    - ftagTruthOriginLabel
+    - ftagTruthVertexIndex
   tracks_loose:
-    - truthOriginLabel
-    - truthVertexIndex
+    - ftagTruthOriginLabel
+    - ftagTruthVertexIndex
 
 track_weight_variables:
   tracks:
-    - truthOriginLabel
-    - truthVertexIndex
+    - ftagTruthOriginLabel
+    - ftagTruthVertexIndex
   tracks_loose:
-    - truthOriginLabel
-    - truthVertexIndex
+    - ftagTruthOriginLabel
+    - ftagTruthVertexIndex
