+++ umami/configs/GNN_Variables.yaml
@@ -43,18 +43,18 @@ track_train_variables:
 
 track_truth_variables:
   tracks:
-    - truthOriginLabel
-    - truthVertexIndex
+    - ftagTruthOriginLabel
+    - ftagTruthVertexIndex
   tracks_loose:
-    - truthOriginLabel
-    - truthVertexIndex
+    - ftagTruthOriginLabel
+    - ftagTruthVertexIndex
 
 track_weight_variables:
   tracks:
-    - truthOriginLabel
-    - truthVertexIndex
+    - ftagTruthOriginLabel
+    - ftagTruthVertexIndex
   tracks_loose:
-    - truthOriginLabel
-    - truthVertexIndex
+    - ftagTruthOriginLabel
+    - ftagTruthVertexIndex
 
 custom_defaults_vars:
