+++ requirements.txt
@@ -1,15 +1,16 @@
-atlas-ftag-tools==0.1.9
+atlas-ftag-tools==0.1.10
 h5py==3.7.0
 matplotlib==3.5.1
 mlxtend==0.19.0
 numpy==1.22.0
 pandas==1.3.5
-puma-hep==0.2.8
+puma-hep==0.3.0
 pydash==5.1.0
-pyyaml==6.0
+pyyaml==6.0.1
 ruamel.yaml==0.17.21
+scipy==1.10.1
 seaborn==0.11.2
 shap==0.40.0
 tables==3.7.0
 tqdm==4.62.3
-umami-preprocessing==0.0.4
+umami-preprocessing==0.0.5
