### 2023-11-07 #5833 @matthewfeickert
This is very delayed, so my apologies, but thank you for your reviews @jpata and @hqucms and thank you also for opening up GitHub Issues on https://github.com/umami-hep/umami/issues. :+1:

@philippgadow Can you please update us on the timeline to address the reviewer Issues and feedback? There doesn't seem to be any activity so far on the GitHub Issues that have been opened. As it has now been a month since you noted these would be addressed in https://github.com/openjournals/joss-reviews/issues/5833#issuecomment-1752803138 it would be good to have an update on where the author team stands.
