+++ umami/preprocessing.py
@@ -1,10 +1,14 @@
 #!/usr/bin/env python
 """Execution script to run preprocessing steps."""
 import argparse
+import os
 from pathlib import Path
+
+from upp.classes.preprocessing_config import PreprocessingConfig
 
 import umami.preprocessing_tools as upt
 from umami.configuration import logger, set_log_level
+from umami.preprocessing_tools.configuration import GeneralSettings
 
 
 def get_parser():
@@ -40,6 +44,14 @@ def get_parser():
         default=int(1e6),
         help="Set the chunk size of the generators.",
     )
+    parser.add_argument(
+        "--upp",
+        default=None,
+        help=(
+            "Specify if UPP should be used for this"
+            "preprocessing config. 1 for True, 0 for False."
+        ),
+    )
 
     # possible job options for the different preprocessing steps
     action = parser.add_mutually_exclusive_group(required=True)
@@ -49,12 +61,14 @@ def get_parser():
         action="store_true",
         help="Prepares hybrid sample (choose sample type).",
     )
+
     action.add_argument(
         "-u",
         "--resampling",
         action="store_true",
         help="Runs resampling.",
     )
+
     action.add_argument(
         "--weighting",
         action="store_true",
@@ -119,65 +133,101 @@ if __name__ == "__main__":
         set_log_level(logger, "DEBUG")
 
     # Load preprocess config file
-    config = upt.PreprocessConfiguration(args.config_file)
+    if args.upp is not None:
+        if args.upp == "0":
+            USE_UPP = False
+            config = upt.PreprocessConfiguration(yaml_config=args.config_file)
+        elif args.upp == "1":
+            USE_UPP = True
+            config = PreprocessingConfig.from_file(
+                Path(args.config_file), split="train"
+            )
+            genral_dict = config.get_umami_general()
+            general = GeneralSettings(**genral_dict)
+            config.mimic_umami_config(general)
+        else:
+            raise ValueError(
+                "You need to define if UPP should be used for this"
+                "preprocessing config. 1 for True, 0 for False."
+            )
+    else:
+        try:
+            config = upt.PreprocessConfiguration(args.config_file)
+            USE_UPP = False
+        except TypeError:
+            config = PreprocessingConfig.from_file(
+                Path(args.config_file), split="train"
+            )
+            genral_dict = config.get_umami_general()
+            general = GeneralSettings(**genral_dict)
+            config.mimic_umami_config(general)
+            USE_UPP = True
 
     # Check for preparation
     if args.prepare:
-        # Copy config to output directory
-        config.copy_to_out_dir("preparation")
-
-        # Check if one specific sample is given
-        if args.sample:
-            samples = [args.sample]
-        # If no specific sample is given iterate over the samples defined in the config
+        if USE_UPP:
+            logger.error("UPP does not support preparation!")
         else:
-            samples = config.preparation.samples.keys()
-            logger.warning(
-                "No --sample was selected, using all in config file! This can"
-                " take a lot of time!"
-            )
-        for iter_sample in samples:
-            # Set the argument in args to this sample and run prepare
-            args.sample = iter_sample
-            preparation_tool = upt.PrepareSamples(args, config)
-            preparation_tool.run()
+            # Copy config to output directory
+            config.copy_to_out_dir("preparation")
+
+            # Check if one specific sample is given
+            if args.sample:
+                samples = [args.sample]
+            # If no specific sample is given iterate
+            # over the samples defined in the config
+            else:
+                samples = config.preparation.samples.keys()
+                logger.warning(
+                    "No --sample was selected, using all in config file! This can"
+                    " take a lot of time!"
+                )
+            for iter_sample in samples:
[…]
