### 2023-11-22 e20511d
Improved documentation + umami paper draft
