+++ docs/plotting/index.md
@@ -1,6 +1,6 @@
 # Plotting in Umami
 
 The umami framework provides extensive plotting options for variable and performance plots.
-Recently also a python API is being developed for an easier use within a python script istead of using the high-level functions and configuration files.
+It serves as a high-level wrapper to the [Plotting UMami Api (`puma`)](https://github.com/umami-hep/puma/).
 
-In the following section you can find more details about the plotting.
+The python API `puma` can also be used stand-alone. In the following section you can find more details about the plotting using Umami to steer the output via configuration files.
