+++ examples/tutorial_jetclass/Variables.yaml
@@ -1,0 +1,7 @@
+label: R10TruthLabel_R22v1
+train_variables:
+  JetKinematics:
+    - pt
+    - eta
+  JetTrainVar:
+    - mass

[…]
