### 2023-11-22 umami-hep/umami#1 @philippgadow
Dear @jpata,

thank you very much for your careful review of the software, its documentation and the associated paper draft.
We apologise for our long delay in answering the comments.

The new version of Umami has much improved documentation which hopefully should address the issues you raised.
We are grateful for the issues, in hindsight we should have provided a tutorial / worked out example how to operate on non-ATLAS data from the start as this will make the software more useful for users outside of the ATLAS collaboration.

I have answered the related issues you raised and closed those where I had the impression that the issue and response were clear to not require discussion. If this assumption was not justified, please re-open the issues and I'd be happy to re-iterate.

Cheers,
Philipp for the Umami developers
