### 2023-11-22 #5833 @philippgadow
Dear @matthewfeickert, dear @jpata, dear @hqucms,

thank you very much for your careful review and the issues you raised. Having separate issues for the different suggested improvements was very useful to follow up on them.
Our apologies that we were so slow in addressing the comments. We made a larger overhaul of the Umami preprocessing and have now improved documentation which most importantly features a fully worked out example how to process the JetClass dataset.

We answered the issues @jpata raised in the project:
https://github.com/umami-hep/umami/issues/1

Kind regards,
Philipp for the Umami developers
