### 2023-11-28 f7b49eb
Add dark mode to mkdocs documentation
