### 2023-12-07 #5833 @philippgadow
Dear @hqucms @jpata @matthewfeickert,

do you find that the modifications to the Umami documentation, most importantly the tutorial how to run the JetClass dataset, sufficient to answer the issues you raised?

Kind regards,
Philipp
