### 2024-01-15 #5833 @philippgadow
Hello @matthewfeickert @jpata @hqucms,

a friendly reminder about the updates we made to the repository, paper draft and documentation. Do the changes answer the points you raised in your review?

Cheers,
Philipp
