### 2024-01-24 #5833 @jpata
Apologies for the delay due to the holiday period and other tasks piling up.

My comments were addressed in the issues I created at https://github.com/openjournals/joss-reviews/issues/5833#issuecomment-1759644622.

I can give a final read to the paper. Should I run `editorialbot generate pdf` to make sure I'm looking at the latest version, or is this something only the editor should do?

Sorry, I forgot.
