### 2024-02-06 #5833 @matthewfeickert
Thank you both very much for your reviews @jpata and @hqucms!

@philippgadow I'll start the editoral pass this week and we can hopefully conclude things in the near future. :+1: We can ignore the missing DOI suggestion by @editorialbot as https://doi.org/10.1163/1574-9347_dnp_e612900 is not the Keras ML framework.
