### 2024-02-07 #5833 @philippgadow
Dear @matthewfeickert ,

thank you very much for starting the editorial pass! It sounds good to me to continue discussion in GitHub issues in the framework's project repository. I am happy to shorten the paper significantly.

> The paper should be between 250-1000 words

We acknowledge that (after the internal review in the ATLAS collaboration), the exceeds the 1000 word mark, so we are happy to discuss in the Github issue where best to drop content from the paper draft to meet the requirements.

Cheers
Philipp
