+++ paper/paper.md
@@ -127,87 +127,26 @@ The application of machine learning in high energy physics, particularly for the
 
 # Development Notes
 
-The development of the package adheres to PEP8 standards [@PEP8:2001]. They are enforced by a continuous integration pipeline in a GitLab project, using the `flake8` linter [@flake8:2023] and the `black` command-line tool for code formatting [@black:2023]. The code quality is tested as part of the continuous integration pipeline with the `pytest` module [@pytest:2004], using unit tests and integration tests.
+The development of the package adheres to PEP8 standards [@PEP8:2001]. The code quality is tested as part of a continuous integration pipeline with the `pytest` module [@pytest:2004], using unit tests and integration tests.
 Documentation of the software is built automatically with the `mkdocs` [@mkdocs:2023] and `sphinx` [@sphinx:2023] modules and deployed to the website [`https://umami.docs.cern.ch`](https://umami.docs.cern.ch).
 The `Umami` toolkit has been released as open-source software under the Apache v2 license. 
 
 # Software description
 
-The `Umami` toolkit provides an integrated workflow including input data preprocessing, algorithm training, and performance evaluation. Furthermore, it interfaces to `lwtnn` [@Guest:2022] to export the trained models to `json` files for `C++` deployment in the ATLAS software stack [@ATLAS:2021].
-
-## Preprocessing
+The `Umami` toolkit provides an integrated workflow including input data preprocessing, algorithm training, and performance evaluation. 
 
 The algorithms are trained on simulated physics processes which provide jets originating from bottom and charm quarks, as well as the background processes which produce jets originating from other sources, such as light-flavour quarks, gluons, or hadronically decaying tau leptons. The input features to the algorithm provide discrimination between the processes. A more detailled discussion of the input features for jet flavour tagging is provided in Ref. [@ATLAS:2019]. The preprocessing in `Umami` addresses several challenges provided both by the nature of the training datasets and the input features.
 
-The steps involved in the preprocessing workflow are illustrated in \autoref{fig:preprocessing}.
+Using `Umami` is not limited to jet flavour tagging but provides support for a broad range of applications. The preprocessing capabilities are demonstrated with simulated physics processes from the JetClass dataset [@JetClass:2022] to distinguish jets originating from Higgs boson decays from jets originating from top quark decays. This represents a similar but slightly different use of machine learning algorithms for jet classification. The software is flexible enough to address this task with only minimal modifications in configuration files.
 
-![Illustration of the preprocessing workflow in `Umami`. Input files with simulated physics processes undergo several stages to provide a training sample, as well as a `json` file with scaling and shifting information ("Scale Dict") and validation/testing samples. These stages include a Preparation stage to define the target classes for the training, a Resampling stage to balance the classes, a Scaling/Shifting stage to transform the range of variables used for training, and a Writing stage to output the resulting samples. The validation/testing samples can also undergo the same resampling as the training sample (not shown).\label{fig:preprocessing}](preprocessing.png){ width=60% }
-
-Typically, the three classes "b-jets" (originating from bottom quarks), "c-jets" (originating from charm quarks), and "light-flavour jets" (originating from gluons and light-flavour quarks) are considered. Several datasets with different physics processes can be combined to a hybrid sample, which is populated over a large jet momentum range.
-The classes in the input dataset are highly imbalanced because the physics processes will predominantly produce light-flavoured jets instead of b-jets or c-jets. 
-Equalising also the distributions for all classes for a certain set of features allows for the parameterisation of the algorithm performance in terms of these features, which can be desirable.
[…]
