### 2024-02-29 #5833 @philippgadow
Dear @matthewfeickert ,

I followed your suggestions to shorten the paper draft significantly and just pushed a change to the `paper` branch of the project. The suggested new draft is here with about 1000 words. I removed parts to adhere more closely to the JOSS guidelines.
[paper.pdf](https://github.com/openjournals/joss-reviews/files/14448517/paper.pdf)

Cheers,
Philipp
