### 2024-03-04 4f70727
Moving tutorial from FTAG documentation
