### 2024-03-07 umami-hep/umami#5 @matthewfeickert
[JOSS Review]: Suggested paper revisions
These are some comments to help the paper component of https://github.com/openjournals/joss-reviews/issues/5833 be more compliant with a JOSS paper. @philippgadow's revisions have already improved things quite a bit, so great job there. :+1:

* [ ] In the ["What should my paper contain?"](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) JOSS docs it notes that the paper should include

> A summary describing the high-level functionality and purpose of the software for a diverse, _non-specialist_ audience.

However, the first setnance of the paper

https://github.com/umami-hep/umami/blob/995170a9378a739be425248d08a9360775aa9699/paper/paper.md?plain=1#L111

mentions jets with no introduction. Non-speacialist here means people outside of particle physics, and so a description of what a jet is minimally required (though it can and should be extremeley brief). While you don't need to also go into great detail as to what the Standard Model there should be some (few word) mention.

* [ ] In

https://github.com/umami-hep/umami/blob/995170a9378a739be425248d08a9360775aa9699/paper/paper.md?plain=1#L121

the

```
It is deployed as a Python module which can be installed with `setuptools` [@setuptools:2023] or used via Docker images [@Merkel:2014].
```

part should be reviesed for accuracy to something like

```
It is available as a Python application and is also distributed via Linux container images.
```

as:
* `umami` currently isn't deployed anywhere as it isn't packaged properly (I'll open a follow up issue on that after this in a few hours) and available on an public Python package index like (the default) PyPI or an alternative.
* `umami` relies heavily on an [(incomplete) `requirements.txt`](https://github.com/umami-hep/umami/blob/e19017383d1385978d7926b036e284f7602fee30/requirements.txt) with exact version pins, making it seem that the intent is to have `umami` be a [Python application and not a library](https://iscinumpy.dev/post/app-vs-library/).
* The use of `setuptools` as a [build backend](https://packaging.python.org/en/latest/tutorials/packaging-projects/#choosing-a-build-backend) (`setuptools` is not an installer) isn't something that needs to be brought up as a user should not ever need to care about how the developer's choice of build backend.
* Linux container images _are_ distributed as Docker images on Docker Hub under the `btagging/umami` image repository and as Singularity images on CVMFS

* [ ] The entire "Development Notes" section

https://github.com/umami-hep/umami/blob/995170a9378a739be425248d08a9360775aa9699/paper/paper.md?plain=1#L128-L132

should be removed as it isn't relevant for the paper.

* [ ] In

https://github.com/umami-hep/umami/blob/995170a9378a739be425248d08a9360775aa9699/paper/paper.md?plain=1#L149

these are libraries, not modules.

* [ ] In https://github.com/openjournals/joss-reviews/issues/5833#issuecomment-1983038931 there are a number of DOIs missing, but flagged bibliographic items like

https://github.com/umami-hep/umami/blob/995170a9378a739be425248d08a9360775aa9699/paper/paper.bib#L25-L32

are also not cited anywhere in the paper and so can just be removed. Other like

https://github.com/umami-hep/umami/blob/995170a9378a739be425248d08a9360775aa9699/paper/paper.bib#L358-L370

that don't have a DOI (somewhat surprisingly) can be safely ignored.

So just a few things left on the paper, but looking very much improved! :+1:
