### 2024-03-07 #5833 @matthewfeickert
@jpata, @hqucms I'm going to need to recall you to ask some questions that came up while I was writing up https://github.com/umami-hep/umami/issues/5. I will also open up a GitHub issue on `umami` later, but you both checked off that the following were complete in your reviews

> * Installation: Does installation proceed as outlined in the documentation?
> * Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

Can you elaborate a bit on how these were tested? In the [Submitting a paper to JOSS docs](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) it is noted that

> In addition, JOSS requires that software should be feature-complete (i.e., no half-baked solutions), packaged appropriately according to common community standards for the programming language being used (e.g., [Python](https://packaging.python.org/), [R](https://r-pkgs.org/index.html)), and designed for maintainable extension (not one-off modifications of existing tools).

which means that for Python packages they should be distributed via a package index (ideally PyPI, or conda-forge if there are complicated external binaries) and are approriately packaged such that they can be installed with their runtime dependencies. The [`umami` installation docs](https://umami.docs.cern.ch/setup/installation/#cloning-the-repository)

(https://github.com/umami-hep/umami/blob/e19017383d1385978d7926b036e284f7602fee30/docs/setup/installation.md?plain=1#L219-L253)

seem to make it clear that the project is missing [the `dependncy` metadata](https://setuptools.pypa.io/en/latest/userguide/dependency_management.html#declaring-required-dependency) to define the actual requirements, and even for an application view the [`requirements.txt`](https://github.com/umami-hep/umami/blob/e19017383d1385978d7926b036e284f7602fee30/requirements.txt) is missing `tensorflow`.

While the reviewer checklist does mention "the documentation", if you've only tested `umami` using provided Docker containers then this is going to require some revision.

---

I'm going to raise an Issue in JOSS that the [Submitting a paper to JOSS docs](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) docs be more explicit about what is required.
