### 2024-03-07 #5833 @jpata
> Can you elaborate a bit on how these were tested?

I used the official docker image as specified in the umami documentation and had no problems with it. I didn't try to pip install the package from PyPI.
