### 2024-03-07 314f991
answer editorial comments and clean up bibliography
