+++ paper/paper.bib
@@ -22,180 +22,6 @@
     primaryClass   = "hep-ex",
 }
 
-% ATLAS software
-@Booklet{ATLAS:2021,
-    author         = "{ATLAS Collaboration}",
-    title          = "{The ATLAS Collaboration Software and Firmware}",
-    howpublished   = "{ATL-SOFT-PUB-2021-001}",
-    url            = "https://cds.cern.ch/record/2767187",
-    year           = "2021",
-}
-
-% Python
-@book{Rossum:2009, 
- author = {Van Rossum, Guido and Drake, Fred L.}, 
- title = {Python 3 Reference Manual}, 
- year = {2009}, 
- isbn = {1441412697}, 
- publisher = {CreateSpace}, 
- address = {Scotts Valley, CA} 
-}
-
-% PEP8
-@techreport{PEP8:2001,
-  author  = {Guido van Rossum and Barry Warsaw and Nick Coghlan},
-  title   = {Style Guide for {Python} Code},
-  year    = {2001},
-  type    = {PEP},
-  number  = {8},
-  url     = {https://www.python.org/dev/peps/pep-0008/},
-}
-
-% YAML
-@misc{YAML:2021,
-  title = {{YAML} Ain’t Markup Language (YAML™) version 1.2},
-  howpublished = {\url{https://yaml.org/spec/1.2.2/}},
-  year=2001,
-  note = {Accessed: 2023-05-11}
-}
-
-% Setuptools
-@misc{setuptools:2023,
-  title = {{Setuptools}},
-  howpublished = {\url{https://github.com/pypa/setuptools}},
-  year=2013,
-  note = {Accessed: 2023-05-11}
-}
-
-
-% Flake8
-@misc{flake8:2023,
-  title = {{Flake8}},
-  howpublished = {\url{https://github.com/PyCQA/flake8}},
-  year=2010,
-  note = {Accessed: 2023-05-11}
-}
-
-% Black
-@misc{black:2023,
-  title = {{Black}},
-  howpublished = {\url{https://github.com/psf/black}},
-  year=2018,
-  note = {Accessed: 2023-05-11}
-}
-
-% Pytest
-@misc{pytest:2004,
-  title =        {pytest 7.3},
-  author = {Krekel, Holger and Oliveira, Bruno and Pfannschmidt, Ronny and Bruynooghe, Floris and Laugher, Brianna and Bruhin, Florian},
-  year =         {2004},
-  url = {https://github.com/pytest-dev/pytest},
-}
-
-% mkdocs
-@misc{mkdocs:2023,
-  title = {{MkDocs}},
-  howpublished = {\url{https://github.com/mkdocs/mkdocs}},
-  year=2014,
-  note = {Accessed: 2023-05-11}
-}
-
-% Pytest
-@misc{sphinx:2023,
-  title =        {Sphinx},
-  author = {Brandl, Georg},
-  year =         {2008},
-  url = {https://www.sphinx-doc.org},
-}
-
-
-% Docker
-@article{Merkel:2014, 
-  title={Docker: lightweight linux containers for consistent development and deployment}, 
-  author={Merkel, Dirk}, 
-  journal={Linux journal}, 
-  volume={2014}, 
-  number={239}, 
-  pages={2}, 
-  year={2014} 
-}
-
-% TensorFlow
-@misc{tensorflow:2015,
-title={ {TensorFlow}: Large-Scale Machine Learning on Heterogeneous Systems},
-url={https://www.tensorflow.org/},
-note={Software available from tensorflow.org},
-author={
-    Mart\'{i}n~Abadi and
-    Ashish~Agarwal and
-    Paul~Barham and
[…]
