### 2024-03-07 umami-hep/umami#5 @philippgadow
Dear @matthewfeickert,

Thank you for taking the time for the thorough editorial review and comments.
Please find the answers here:

- request: add explanation of what a jet is
    - answer: we rephrased the first sentence (additions marked in italics)

    "Flavour-tagging, the identification of *collimated sprays of particles ("jets")* originating from bottom and charm quarks, is a critically important technique in the data analysis of the ATLAS experiment [@ATLAS:2008] at the Large Hadron Collider [@Evans:2008]."

- request: change "It is deployed as a Python module which can be installed with `setuptools` [@setuptools:2023] or used via Docker images [@Merkel:2014]." to "It is available as a Python application and is also distributed via Linux container images."
    - answer: Done

- request: remove "Development notes" section
    - answer: Done

- request: change " The performance of the chosen model can be evaluated in publication-grade plots, which are steered with configuration files. The plots are created using the `matplotlib` [@Hunter:2007] and `puma` [@Birk:2023] Python modules." to " The performance of the chosen model can be evaluated in publication-grade plots, which are steered with configuration files. The plots are created using the `matplotlib` [@Hunter:2007] and `puma` [@Birk:2023] Python libraries."
    - answer: Done

- request: clean bibliography
    - answer: Done

Please close the comment and mark the checkboxes if you find our responses satisfactory. If not, we are happy to iterate here.
