### 2024-03-08 #5833 @matthewfeickert
> I didn't try to pip install the package

@jpata @hqucms Hm. Thanks for this information. I'll discuss this with the JOSS Editorial Board this week to see how to proceed here.

> from PyPI.

Though note that there is no PyPI package (which is part of the concern) as https://pypi.org/project/umami is a different project.
