### 2024-03-08 umami-hep/umami#5 @matthewfeickert
Thanks for the fast turn around @philippgadow! From this latest build (https://github.com/openjournals/joss-reviews/issues/5833#issuecomment-1985338649) I think this is looking very good.

One final revision is that in

https://github.com/umami-hep/umami/blob/314f991e499a50b5bf04de8849546cd2b39f193f/paper/paper.md?plain=1#L110

```
It is applied in precision measurements of the Standard Model, e.g. in characterisations of the Higgs boson properties, as well as in searches for yet unknown phenomena.
```

introduces the Standard Model, but non-physicsts won't know what that is. So if you put in a similar level explanation as you did for jets (e.g. less than a sentence) that would be great.

After that I think the paper is ready to go. :+1:
