### 2024-03-08 umami-hep/umami#5 @philippgadow
Hi @matthewfeickert !

I think that if readers don't know what the Standard Model is, then they also will not know what a Higgs boson or a bottom or charm quark are. However, I see the point of making the paper accessible to a non-specialist audience, so I rephrased as follows:

The sentence

> It is applied in precision measurements of the Standard Model, e.g. in characterisations of the Higgs boson properties, as well as in searches for yet unknown phenomena.

has been rephrased to

> It is applied in precision measurements of the Standard Model of particle physics,
which is the theory describing fundamental forces and classifying all known elementary particles, as well as in searches for yet unknown phenomena.

I appreciate the level of detail you take in the review, thank you!
