### 2024-05-21 umami-hep/umami#6 @matthewfeickert
:wave: @philippgadow I'm just checking in on this Issue for https://github.com/openjournals/joss-reviews/issues/5833 not to rush you and the team or be pushy, but to make it clear that I haven't forgotten about this review and abandoned it. If there are questions that I can provide any recommendations or further input on I am happy to do so.
