### 2024-06-27 #5833 @matthewfeickert
@philippgadow, can you please give us an update on Issue https://github.com/umami-hep/umami/issues/6?
