### 2024-07-16 umami-hep/umami#6 @afroch
Hi @matthewfeickert. Sorry for the long delay, but I'm currently looking into your comments from above. I adapted already your bullet points 3 and 4 but I have a question concerning your first and second bullet point:

1: When I understand this correctly, the TODO here is to update the `pyproject.toml` with the little code snippet you provided and add the requirements from the `requirements.txt` file to the `setup.cfg`. Is this correct? If so, I saw there is a possibility to do this dynamically. I guess this would be good?

2: Which packages do you talk about when writing "correct application run requirements with deploy/run versions pinned"? From my point, I think we pinned down all the different packages that we use with the version required. Could you clarify what you meant with that?
