### 2024-09-17 umami-hep/umami#6 @matthewfeickert
Sorry for _my_ long delay, @afroch, and the inconveniences it has caused.

> I adapted already your bullet points 3 and 4

Fantastic and thank you. Do you happen to have these on a branch of https://github.com/umami-hep/umami/ or a public fork that I can look at and give feedback on? Feel free to open PRs to address them and tag me there too.

> 1: When I understand this correctly, the TODO here is to update the pyproject.toml with the little code snippet you provided

Correct. The build-system metadata is incomplete at the moment

https://github.com/umami-hep/umami/blob/c7e7d164f554f07a20dcfbf760a47c828e638abc/pyproject.toml#L1-L2

and so not compliant with [PEP 518](https://peps.python.org/pep-0518/), which will be a problem in the future for installers. All that is required here is to correctly specify the `build-backend` metadata and update the `requires`. So you can just update the `[build-system]` to be

```toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
```

c.f. The [Scientific Python library development guide on simple packaging](https://learn.scientific-python.org/development/guides/packaging-simple/#pyprojecttoml-build-system) for more details.

> and add the requirements from the requirements.txt file to the setup.cfg. Is this correct?

`umami` should be installable with standard tooling (e.g. `python -m pip install` or `uv pip install`) without needing any additional files, and so having the runtime requirements be specified in the `setup.cfg` is necessary. Though (going into question 2 a bit) the known lower bounds (`>=`) of the dependencies should be specified (unless there is a known breaking change in a requirement which necessitates an upper bound be added `>=X,<Y` or a pin `==Y`).

> In short, this means that for libraries all the required runtime dependencies need to be included in the packaging metadata with known lower bounds (if you don't want to empirically determine these lower bounds it is okay to select reasonable known working versions) so the installer knows what versions are required. For a Python application, in addition to the library requirements, a `requirements.txt` or (preferably) a lock file (created by some well established tool like `pip-tools`, `conda-lock`, `poetry`, `uv` as there is [currently no formal Python lock file standard](https://discuss.python.org/t/lock-files-again-but-this-time-w-sdists/46593)) should also be provided.

(I'm copying that from https://github.com/umami-hep/umami/issues/6#issue-2181036198 to avoid you having to reread that all again)

> If so, I saw there is a possibility to do this dynamically. I guess this would be good?

Sorry, can you elaborate more on what "do this dynamically" means?

> 2: Which packages do you talk about when writing "correct application run requirements with deploy/run versions pinned"? From my point, I think we pinned down all the different packages that we use with the version required. Could you clarify what you meant with that?

When answering this question it helps to have a clear understanding of what the [difference between a library and an application is](https://iscinumpy.dev/post/app-vs-library/). As it was agreed that `umami` is not intended to be used as a library, but as a Python application, it is then important to _fully specify_ the runtime environment. The contents of the `requirements.txt`

https://github.com/umami-hep/umami/blob/c7e7d164f554f07a20dcfbf760a47c828e638abc/requirements.txt#L1-L16

do not do this as they only specify the versions of the high level requirements &mdash; they do not specify the versions of those requirements' dependencies at all. A practical way to see this it so generate a lock file from the `requirements.txt` with `uv` and note that number of dependencies is quite different (there are 25 dependencies that are not being specified and could very well break any and all behavior in future updates):

```console
$ docker run --rm -ti python:3.12 /bin/bash
root@666372a306f4:/# python -m venv venv && . venv/bin/activate
(venv) root@666372a306f4:/# python -m pip --quiet install --upgrade uv
(venv) root@666372a306f4:/# curl -sLO https://raw.githubusercontent.com/umami-hep/umami/c7e7d164f554f07a20dcfbf760a47c828e638abc/requirements.txt
(venv) root@666372a306f4:/# uv pip compile --generate-hashes --output-file=requirements.lock requirements.txt
(venv) root@666372a306f4:/# grep "==" requirements.txt | wc -l
16
(venv) root@666372a306f4:/# grep "==" requirements.lock | wc -l
41
```

An application should have the runtime environment fully specified with a lock file built from a high level summary of the direct retirements. The existing `requirements.txt` is an example of such a high level specification.

---

If you have additional questions please feel free to ask for clarification here. I'm happy to discuss anything. As noted in email communication I'm currently a bit occupied for personal/professional reasons, but with this now resurfaced in GitHub Issues I can be substantially more responsive to help you finish this JOSS submission.
