### 2024-09-17 umami-hep/umami#6 @afroch
Hi @matthewfeickert. Thanks for the very nice answer! I tried to adapt all the comments you had and cumulated them in the following MR (Gitlab) [!752](https://gitlab.cern.ch/atlas-flavor-tagging-tools/algorithms/umami/-/merge_requests/752). The project should be visible to everyone and once it is merged into master, the github instance of umami will be updated by a cron job.

Feel free to drop any comments on the MR or here.
