### 2024-09-17 umami-hep/umami#6 @matthewfeickert
(I'll double post here and on CERN's GitLab)

@afroch MR https://gitlab.cern.ch/atlas-flavor-tagging-tools/algorithms/umami/-/merge_requests/752 is sitll in draft, so maybe it isn't done yet, but from https://github.com/umami-hep/umami/issues/6#issuecomment-2354707896 you still need to define the runtime requirements in the metadata (this is `install_requires` for `setuptools` and `dependencies` for all other build backends).

Aside: (This is a comment NOT a requirement to proceed) You don't use any `setuptools` particular features here, so you could also migrate this all to `hatchling` and simplify everything to exist in `pyproject.toml`. `hatch` does this for you automatically with `hatch new --init`

```console
$ docker run --rm -ti python:3.12 /bin/bash
root@88dde1034a68:/# python -m venv venv && . venv/bin/activate
(venv) root@88dde1034a68:/# git clone --recursive https://github.com/umami-hep/umami.git
Cloning into 'umami'...
remote: Enumerating objects: 11736, done.
remote: Total 11736 (delta 466), reused 792 (delta 433), pack-reused 10907 (from 1)
Resolving deltas: 100% (7807/7807), done.
(venv) root@88dde1034a68:/# cd umami/
(venv) root@88dde1034a68:/umami# python -m pip --quiet install --upgrade hatch
(venv) root@88dde1034a68:/umami# hatch new --init
(venv) root@88dde1034a68:/umami# git diff
diff --git a/pyproject.toml b/pyproject.toml
index 11d1e97..caadc7b 100644
--- a/pyproject.toml
+++ b/pyproject.toml
@@ -1,5 +1,31 @@
 [build-system]
-requires = ["setuptools", "wheel"]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[project]
+name = "umami"
+dynamic = ["version"]
+description = "Machine learning based flavour tagging training framework."
+readme = "README.md"
+license = ""
+requires-python = ">= 3.8"
+keywords = [
+    "flavour",
+    "learning",
+    "machine",
+    "tagging",
+]
+
+[project.urls]
+Homepage = "https://gitlab.cern.ch/atlas-flavor-tagging-tools/algorithms/umami"
+
+[tool.hatch.version]
+path = "umami/__init__.py"
+
+[tool.hatch.build.targets.sdist]
+include = [
+    "/umami",
+]

 [tool.black]
 line-length = 88
```

Also do not try to make `requirements.txt` a self created lock file. This is an anti-goal and now removes any benefits of understanding what your high level dependencies are. Keep a high level runtime requirements file separate from a **tool** built lock file. If the project runtime requirements are properly defined you can generate a lock file from them like in the following example:

```console
$ docker run --rm -ti python:3.12 /bin/bash
root@49d9431484e0:/# python -m venv venv && . venv/bin/activate
(venv) root@49d9431484e0:/# python -m pip --quiet install --upgrade uv
(venv) root@49d9431484e0:/# git clone --recursive https://github.com/scikit-hep/awkward.git
(venv) root@49d9431484e0:/# cd awkward/
(venv) root@49d9431484e0:/# uv pip compile --generate-hashes --output-file=requirements.txt pyproject.toml
```

This example demos how you could keep the `requirements.txt` name without having to teach people about `requirements.lock` and also benefit from having your requirements defined properly.
