+++ docker/umami-slim/Dockerfile
@@ -2,7 +2,7 @@
 # dependence
 FROM ${BASE_IMAGE}
 
-## ensure locale is set during build
+# Ensure locale is set during build
 ENV LANG C.UTF-8
 ENV LC_ALL C.UTF-8
 
@@ -15,6 +15,9 @@ RUN apt-get update && \
 
 COPY . /umami
 WORKDIR /umami
+
+# Remove tensorflow from requirements list
+RUN sed -i '/tensorflow/d' requirements.txt
 
 RUN pip install -r requirements.txt
 RUN cd /umami && python -m pip install .
