+++ setup.cfg
@@ -9,6 +9,26 @@ url = https://gitlab.cern.ch/atlas-flavor-tagging-tools/algorithms/umami
 zip_safe = False
 include_package_data = True
 python_requires = >= 3.8
+install_requires =
+    atlas-ftag-tools >= 0.1.10
+    atlasify >= 0.8.0
+    h5py >= 3.7.0
+    matplotlib >= 3.5.1, < 3.5.2
+    mlxtend >= 0.19.0
+    numpy >= 1.22.0
+    pandas >= 1.3.5, < 1.3.6
+    puma-hep >= 0.3.0, < 0.3.1
+    pydash >= 5.1.0
+    pyyaml >= 6.0.1
+    ruamel-yaml >= 0.17.21
+    scikit-learn >= 1.0.2
+    scipy >= 1.10.1
+    seaborn >= 0.11.2
+    shap >= 0.40.0
+    tables >= 3.7.0
+    tensorflow >= 2.11.0, < 2.11.1
+    tqdm >= 4.62.3
+    umami-preprocessing >= 0.0.6
 packages = find:
 scripts =
     umami/preprocessing.py
