### 2024-09-30 umami-hep/umami#6 @afroch
Hi @matthewfeickert. The corresponding MR adapting all your suggestions ([!752](https://gitlab.cern.ch/atlas-flavor-tagging-tools/algorithms/umami/-/merge_requests/752) in Gitlab) is now merged and we fixed all the outstanding issues with the deployment part of the pipeline.
