### 2024-10-01 #5833 @matthewfeickert
@philippgadow and @afroch have resolved https://github.com/umami-hep/umami/issues/6, so we can now restart editorial review of the paper and then move to the publication steps.
