### 2024-10-01 umami-hep/umami#5 @matthewfeickert
> [!NOTE]
> Context on the delay between the last communications is here: https://github.com/openjournals/joss-reviews/issues/5833#issuecomment-2384778921

---

Okay. While the seciton

https://github.com/umami-hep/umami/blob/fe0f292d16d380c8c2817d87bb4046045f0591c6/paper/paper.md?plain=1#L127

isn't needed, I think it is okay to keep in so that this can move forward towards publication. I'll close this as done now.
