### 2024-10-01 #5833 @matthewfeickert
I'm marking https://github.com/umami-hep/umami/issues/5 as resolved: https://github.com/umami-hep/umami/issues/5#issuecomment-2384803591

---

@philippgadow and @afroch all that's left to do now (as noted in https://github.com/openjournals/joss-reviews/issues/5833#issuecomment-1928940964) before recommending publication is to ensure that there is a long term public archive of the code that was reviewed. In this case the last code release was [`umami` `v0.2.1`](https://github.com/umami-hep/umami/releases/tag/0.21) and there has been [some development between then and the current `HEAD`](https://github.com/umami-hep/umami/compare/0.21...a8a61cc418a465a4583687024450276a58ae4ccc), so we'd request making a new release (as you see fit with respect to SemVer).

We'd suggest depositing the code either with [Zenodo](https://zenodo.org/) or with [figshare](https://figshare.com/) to get an archive with a DOI. If you use Zenodo there is [an (optional) GitHub integration](https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content) that can create a Zenodo archive for you anytime you make a GitHub release of your code.

[![Zenodo_DOI_guide](https://user-images.githubusercontent.com/5142394/234173148-2226325d-0a42-4526-b546-75b469d26c4c.png)](https://zenodo.org/account/settings/github/)

Though as the GitHub source is a sync with the CERN GitLab repository, you may want to use a different approach to interface with Zenodo.

Once you make a new release of `umami` and have a DOI for the archive, and have verified all related author tasks listed in https://github.com/openjournals/joss-reviews/issues/5833#issuecomment-1928940964, please just report them here and I'll have @editorialbot update the version listed in the review and add the archive.

Let me know if you have any questions. :+1:
