### 2024-10-02 #5833 @philippgadow
Dear @matthewfeickert ,

thank you so much for your patience and for making very helpful suggestions. We appreciate you following the process very much, thank you!

We list below the tasks we did

- [x] Double check authors and affiliations (including ORCIDs): **updated author affiliations**
- [x] Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper. **The latest release of Umami is v0.22 and it is linked here: [v0.22](https://github.com/umami-hep/umami/releases/tag/v0.22)**
- [x] Archive the release on Zenodo/figshare/etc and post the DOI here. **Release v0.22 was archived on Zenodo. It has the DOI** [![DOI](https://zenodo.org/badge/622863715.svg)](https://doi.org/10.5281/zenodo.13881189)
