### 2024-10-02 #5833 @matthewfeickert
@philippgadow excellent! :rocket: I'll go ahead and move forward with recommending acceptance for publication, but while I'm doing that can you please edit the metadata of https://doi.org/10.5281/zenodo.13881189 to

> Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.

>

So specifically, change the title from

> umami-hep/umami: v0.22

to

> Umami: A Python toolkit for jet flavour tagging

(Aside: this is a very easy thing to miss, and we only require this for the publication release.)

We can do these things asynchronously though. :+1:
