### 2024-10-02 #5833 @matthewfeickert
@openjournals/pe-eics note https://github.com/openjournals/joss-reviews/issues/5833#issuecomment-2388881468 for archive metadata edits that need to happen.
