### 2024-10-06 #5833 @philippgadow
Dear @matthewfeickert, I confirm that the proofs look correct. The links (except for the JOSS DOI which does not exist yet) all work and I spotted no typos or formatting errors.
