### 2024-10-09 #5833 @matthewfeickert
(I'm once again traveling and so in transit and so delayed, but) congratulations @philippgadow and ATLAS Flavour Tagging on your JOSS publication! Great work! Also many thanks @jpata and @hqucms for your reviews and all of your hard work. It is very much appreciated!
